Preserve Sourcify source selection across entire app
