+++ src/Main.tsx
@@ -1,30 +1,19 @@
-import React, { Suspense, useMemo, useState } from "react";
+import React, { Suspense } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 import { Outlet } from "react-router";
 import Header from "./Header";
 import ErrorFallback from "./components/ErrorFallback";
-import { SourcifySourceName } from "./sourcify/useSourcify";
-import { AppConfig, AppConfigContext } from "./useAppConfig";
 
 const Main: React.FC = () => {
-  const [sourcifySource, setSourcifySource] =
-    useState<SourcifySourceName | null>(null);
-  const appConfig = useMemo((): AppConfig => {
-    return {
-      sourcifySource,
-      setSourcifySource,
-    };
-  }, [sourcifySource, setSourcifySource]);
-
   return (
-    <AppConfigContext.Provider value={appConfig}>
+    <>
       <Header />
       <ErrorBoundary FallbackComponent={ErrorFallback}>
         <Suspense fallback={<div className="w-screen h-screen"></div>}>
           <Outlet />
         </Suspense>
       </ErrorBoundary>
-    </AppConfigContext.Provider>
+    </>
   );
 };
